requirements.txt
flask
openai
python-dotenv
psycopg2-binary
gunicorn
